requirements.txt
fastapi
uvicorn
pydantic
pyjwt
authlib
requests
python-dotenv
boto3
sqlalchemy
psycopg2-binary